refactor[#445] : MarkingItem features -> widgets 레이어로 이동
