+++ src/features/marking/ui/index.ts
@@ -1,4 +1,3 @@
-export * from "./markingItem";
 export * from "./markingFormModal";
 export * from "./markingFilter";
 export * from "./editMarkingFormModal";
