+++ src/widgets/marking/store/index.ts
@@ -1,0 +1,1 @@
+export * from "./markingItem";
