+++ src/widgets/marking/ui/index.ts
@@ -1,2 +1,3 @@
 export * from "./markingThumbnailGrid";
 export * from "./temporaryMarkingItem";
+export * from "./markingItem";
